chore: rename rules to follow eslint
