+++ src/rules/no-template-bind.ts
@@ -1,5 +1,5 @@
 /**
- * @fileoverview Disallows array `.map` in templates
+ * @fileoverview Disallows arrow functions and `.bind` in templates
  * @author James Garbutt <htttps://github.com/43081j>
  */
 
@@ -13,9 +13,9 @@ import * as ESTree from 'estree';
 const rule: Rule.RuleModule = {
   meta: {
     docs: {
-      description: 'Disallows array `.map` in templates',
+      description: 'Disallows arrow functions and `.bind` in templates',
       category: 'Best Practices',
-      url: 'https://github.com/43081j/eslint-plugin-lit/blob/master/docs/rules/template-no-map.md'
+      url: 'https://github.com/43081j/eslint-plugin-lit/blob/master/docs/rules/no-template-bind.md'
     }
   },
 
@@ -25,6 +25,35 @@ const rule: Rule.RuleModule = {
     //----------------------------------------------------------------------
     // Helpers
     //----------------------------------------------------------------------
+
+    /**
+     * Determines whether a node is a disallowed expression
+     * or not.
+     *
+     * @param {ESTree.Node} node
+     * @return {boolean}
+     */
+    function isDisallowedExpr(node: ESTree.Node): boolean {
+      if (node.type === 'ArrowFunctionExpression' ||
+          node.type === 'FunctionExpression') {
+        return true;
+      }
+
+      if (node.type === 'CallExpression' &&
+          node.callee.type === 'MemberExpression' &&
+          node.callee.property.type === 'Identifier' &&
+          node.callee.property.name === 'bind') {
+        return true;
+      }
+
+      if (node.type === 'ConditionalExpression') {
+        return isDisallowedExpr(node.test) ||
+          isDisallowedExpr(node.consequent) ||
+          isDisallowedExpr(node.alternate);
+      }
+
+      return false;
+    }
 
     //----------------------------------------------------------------------
     // Public
@@ -36,13 +65,10 @@ const rule: Rule.RuleModule = {
           node.tag.type === 'Identifier' &&
           node.tag.name === 'html') {
           for (const expr of node.quasi.expressions) {
-            if (expr.type === 'CallExpression' &&
-              expr.callee.type === 'MemberExpression' &&
-              expr.callee.property.type === 'Identifier' &&
-              expr.callee.property.name === 'map') {
+            if (isDisallowedExpr(expr)) {
               context.report({
                 node: expr,
-                message: '`.map` is disallowed in templates, move the expression out of the template instead'
+                message: 'Arrow functions and `.bind` must not be used in templates'
               });
             }
           }
