+++ src/rules/no-useless-template-literals.ts
@@ -1,5 +1,5 @@
 /**
- * @fileoverview Disallows literal values in templates
+ * @fileoverview Disallows redundant literal values in templates
  * @author James Garbutt <htttps://github.com/43081j>
  */
 
@@ -13,10 +13,10 @@ import * as ESTree from 'estree';
 const rule: Rule.RuleModule = {
   meta: {
     docs: {
-      description: 'Disallows literal values in templates',
+      description: 'Disallows redundant literal values in templates',
       category: 'Best Practices',
       recommended: true,
-      url: 'https://github.com/43081j/eslint-plugin-lit/blob/master/docs/rules/template-no-literal.md'
+      url: 'https://github.com/43081j/eslint-plugin-lit/blob/master/docs/rules/no-useless-template-literals.md'
     }
   },
